refactor: (DatePicker) recouple date and week utils
